finished banned screen, create entry for listings, filter listings

+++ rpghub/src/main/java/com/revature/scramble/controller/EntryController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import com.revature.scramble.service.EntryService;
 import com.revature.scramble.service.ListingService;
+import com.revature.scramble.models.Session;
 import com.revature.scramble.repository.ListingDao;
 import com.revature.scramble.repository.entities.Entry;
 import com.revature.scramble.repository.entities.Listing;
@@ -13,20 +14,18 @@ import io.javalin.http.Handler;
 public class EntryController {
 
     public static Handler fetch_post_create_entry = ctx ->{
-        if(HomeController.check_account()){
+        System.out.println("fetch_post_create_entry");
+        if(HomeController.check_account() && !Session.is_frozen){
             //Log 
             //database request
-            int entry_id = Integer.parseInt(ctx.formParam("entry_id"));
-            int list_id = Integer.parseInt(ctx.formParam("list_id"));
-            int user_id = Integer.parseInt(ctx.formParam("user_id"));
-            String user_role = ctx.formParam("user_role");
-            String user_note = ctx.formParam("user_note");
-            String user_status = ctx.formParam("status");
-            Entry new_entry = new Entry(entry_id,list_id,user_id,user_role,user_note,user_status);
-
+            Entry new_entry = new Entry(-1,Integer.parseInt(ctx.formParam("list_id")), Integer.parseInt(ctx.formParam("user_id")), ctx.formParam("user_role"), ctx.formParam("user_note"),"Pending");
+            System.out.println("new_entry");
             //call the service method here
             EntryService.create_new_entry(new_entry);
-            //Entry_Service.create_new_entry()
+            System.out.println("fetch_post_create_entry: creation done");
+            ctx.redirect("/listing/manage/"+ctx.formParam("list_id")+"/view");
+        }else if(Session.is_frozen){
+            ctx.redirect("/frozen");
         }
         else{
             //Log
@@ -36,7 +35,7 @@ public class EntryController {
     };
 
     public static Handler fetch_post_update_entry = ctx ->{
-        if(HomeController.check_account()){
+        if(HomeController.check_account() && !Session.is_frozen){
             //Log 
             //database request
             String user_status = ctx.formParam("status");
@@ -47,6 +46,10 @@ public class EntryController {
             //Update particular listing
             Listing new_listing = new Listing(Integer.parseInt(ctx.formParam("list_id")), Integer.parseInt(ctx.formParam("user_id")), ctx.formParam("list_name"), ctx.formParam("dungeonName"), Integer.parseInt(ctx.formParam("max_size")), Integer.parseInt(ctx.formParam("cur_size")));
             ListingService.update_listing_service(new_listing);
+            
+            ctx.redirect("/listing/"+ctx.formParam("list_id"));
+        }else if(Session.is_frozen){
+            ctx.redirect("/frozen");
         }
         else{
             //Log
@@ -56,11 +59,16 @@ public class EntryController {
     };
 
     public static Handler fetch_post_delete_entry = ctx ->{
-        if(HomeController.check_account()){
-            //Log 
-            //database request
-            int entry_id = Integer.parseInt(ctx.formParam("entry_id"));
-            EntryService.delete_entry(entry_id);
+        if(HomeController.check_account() && !Session.is_frozen){
+            if(Integer.parseInt(ctx.formParam("entry_id")) != -1 && Integer.parseInt(ctx.formParam("list_id")) != -1){
+                //Log 
+                //database request
+                int entry_id = Integer.parseInt(ctx.formParam("entry_id"));
+                EntryService.delete_entry(entry_id);
+            }
+            ctx.redirect("/listing/manage/"+ctx.formParam("list_id")+"/view");
+        }else if(Session.is_frozen){
+            ctx.redirect("/frozen");
         }
         else{
             //Log
@@ -70,13 +78,15 @@ public class EntryController {
     };
 
     public static Handler fetch_get_all_entries = ctx ->{
-        if(HomeController.check_account()){
+        if(HomeController.check_account() && !Session.is_frozen){
             //Log 
             //Render json of entries
             int list_id = Integer.parseInt(ctx.pathParam("list_id"));
             List<Entry> list_of_entry = EntryService.get_all_entries(list_id);
             String json_string = EntryService.get_json_string_from_entries(list_of_entry);
             ctx.json(json_string);
+        }else if(Session.is_frozen){
+            ctx.redirect("/frozen");
         }
         else{
             //Log
